PORT number added as env variable

+++ backend/controllers/recipe.controller.js
@@ -62,10 +62,14 @@ export const updateRecipe = async (req, res) => {
 export const deleteRecipe = async (req, res) => {
   const {id} = req.params;
 
+  if(!mongoose.Types.ObjectId.isValid(id)) {
+    return res.status(404).json({ success: false, message: "Invalid Recipe Id" });
+  }
+
   try {
     await Recipe.findByIdAndDelete(id);
-    res.status(200).json({ success: true, message: "Product deleted" });
+    res.status(200).json({ success: true, message: "Recipe deleted" });
   } catch (error) {
-    res.status(404).json({ success: false, message: "Product not found" });
+    res.status(500).json({ success: false, message: "Server Error" });
   }
 };
